[refactor] 폴더위치 변경
