+++ build.gradle
@@ -3,6 +3,7 @@ plugins {
 	id 'org.springframework.boot' version '3.0.12'
 	id 'io.spring.dependency-management' version '1.1.3'
 	id "org.flywaydb.flyway" version "10.0.0"
+	id "com.ewerk.gradle.plugins.querydsl" version "1.0.10"
 }
 
 group = 'com.project'
@@ -60,10 +61,17 @@ dependencies {
 
 	//batch
 	implementation 'org.springframework.boot:spring-boot-starter-batch'
+
 	//elastic
 	implementation 'org.elasticsearch.client:elasticsearch-rest-high-level-client:7.17.4'
 	implementation 'org.elasticsearch.client:elasticsearch-rest-client:7.17.4'
 	implementation 'org.springframework.boot:spring-boot-starter-data-elasticsearch'
+
+	//querydsl
+	implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
+	annotationProcessor 'com.querydsl:querydsl-apt:5.0.0:jakarta'
+	annotationProcessor "jakarta.annotation:jakarta.annotation-api"
+	annotationProcessor "jakarta.persistence:jakarta.persistence-api"
 
 
 	testImplementation 'org.springframework.batch:spring-batch-test'
@@ -84,3 +92,23 @@ tasks.named('bootBuildImage') {
 tasks.named('test') {
 	useJUnitPlatform()
 }
+
+//querydsl 추가 시작
+def querydslDir = "$buildDir/generated/querydsl"
+querydsl {
+	jpa = true
+	querydslSourcesDir = querydslDir
+}
+sourceSets {
+	main.java.srcDir querydslDir
+}
+configurations {
+	querydsl.extendsFrom compileClasspath
+}
+compileQuerydsl {
+	options.annotationProcessorPath = configurations.querydsl
+}
+clean {
+	delete file(querydslDir)
+}
+//querydsl 추가 끝
