+++ src/main/java/com/project/dogwalker/config/ElasticSearchConfig.java
@@ -1,6 +1,6 @@
 package com.project.dogwalker.config;
 
-import com.project.dogwalker.domain.user.elastic.WalkerSearchRepository;
+import com.project.dogwalker.domain.user.walker.elastic.WalkerSearchRepository;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.elasticsearch.client.ClientConfiguration;
