+++ src/main/java/com/project/dogwalker/domain/user/walker/elastic/WalkerSearchRepository.java
@@ -1,4 +1,4 @@
-package com.project.dogwalker.domain.user.elastic;
+package com.project.dogwalker.domain.user.walker.elastic;
 
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 import org.springframework.stereotype.Repository;
