+++ src/main/java/com/project/dogwalker/member/service/OauthServiceImpl.java
@@ -12,8 +12,8 @@ import com.project.dogwalker.domain.user.User;
 import com.project.dogwalker.domain.user.UserRepository;
 import com.project.dogwalker.domain.user.customer.CustomerDogInfo;
 import com.project.dogwalker.domain.user.customer.CustomerDogInfoRepository;
-import com.project.dogwalker.domain.user.elastic.WalkerDocument;
-import com.project.dogwalker.domain.user.elastic.WalkerSearchRepository;
+import com.project.dogwalker.domain.user.walker.elastic.WalkerDocument;
+import com.project.dogwalker.domain.user.walker.elastic.WalkerSearchRepository;
 import com.project.dogwalker.domain.user.walker.WalkerSchedule;
 import com.project.dogwalker.domain.user.walker.WalkerScheduleRepository;
 import com.project.dogwalker.domain.user.walker.WalkerServicePrice;
