+++ src/test/java/com/project/dogwalker/member/service/OauthServiceImplTest.java
@@ -14,8 +14,8 @@ import com.project.dogwalker.domain.user.Role;
 import com.project.dogwalker.domain.user.User;
 import com.project.dogwalker.domain.user.UserRepository;
 import com.project.dogwalker.domain.user.customer.CustomerDogInfoRepository;
-import com.project.dogwalker.domain.user.elastic.WalkerDocument;
-import com.project.dogwalker.domain.user.elastic.WalkerSearchRepository;
+import com.project.dogwalker.domain.user.walker.elastic.WalkerDocument;
+import com.project.dogwalker.domain.user.walker.elastic.WalkerSearchRepository;
 import com.project.dogwalker.domain.user.walker.WalkerScheduleRepository;
 import com.project.dogwalker.domain.user.walker.WalkerServicePriceRepository;
 import com.project.dogwalker.exception.member.LoginMemberNotFoundException;
